fix a warning
git-svn-id: svn+ssh://svn.cprover.org/srv/svn/cbmc/trunk@4892 6afb6bc1-c8e4-404c-8f48-9ae832c5b171

+++ src/goto-instrument/dump_c.cpp
@@ -611,7 +611,7 @@ Purpose:
 \*******************************************************************/
 
 #define ADD_TO_SYSTEM_LIBRARY(v, header) \
-  for(int i=0; i<sizeof(v)/sizeof(char*); ++i) \
+  for(size_t i=0; i<sizeof(v)/sizeof(char*); ++i) \
     system_library_map.insert( \
       std::make_pair(std::string("c::")+v[i], header))
